walt_restauth_1/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
commonmark==0.9.1
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-extensions==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
httpie==3.2.1
idna==3.3
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
phonenumbers==8.13.11
Pillow==9.2.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
Werkzeug==2.2.1
